Changes to tween component
